performed code cleanup
