+++ pkg/vmo/controller.go
@@ -18,7 +18,7 @@ import (
 	listers "github.com/verrazzano/verrazzano-monitoring-operator/pkg/client/listers/vmcontroller/v1"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/config"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/constants"
-	metricsExporter "github.com/verrazzano/verrazzano-monitoring-operator/pkg/metricsexporter"
+	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/metricsexporter"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/opensearch"
 	dashboards "github.com/verrazzano/verrazzano-monitoring-operator/pkg/opensearch_dashboards"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/signals"
@@ -430,8 +430,8 @@ func (c *Controller) syncHandler(key string) error {
 // with the current status.
 func (c *Controller) syncHandlerStandardMode(vmo *vmcontrollerv1.VerrazzanoMonitoringInstance) error {
 	var errorObserved bool
-	metricsExporter.GetFunctionMetrics(metricsExporter.NamesReconcile).LogStart()
-	defer func() { metricsExporter.GetFunctionMetrics(metricsExporter.NamesReconcile).LogEnd(errorObserved) }()
+	metricsexporter.GetFunctionMetrics(metricsexporter.NamesReconcile).LogStart()
+	defer func() { metricsexporter.GetFunctionMetrics(metricsexporter.NamesReconcile).LogEnd(errorObserved) }()
 
 	originalVMO := vmo.DeepCopy()
 
@@ -521,7 +521,7 @@ func (c *Controller) syncHandlerStandardMode(vmo *vmcontrollerv1.VerrazzanoMonit
 	if !errorObserved {
 		deploymentsDirty, err = CreateDeployments(c, vmo, pvcToAdMap, existingCluster)
 		if err != nil {
-			metricsExporter.GetFunctionMetrics(metricsExporter.NamesReconcile).IncError()
+			metricsexporter.GetFunctionMetrics(metricsexporter.NamesReconcile).IncError()
 			errorObserved = true
 		}
 	}
@@ -542,7 +542,7 @@ func (c *Controller) syncHandlerStandardMode(vmo *vmcontrollerv1.VerrazzanoMonit
 		c.log.Debugf("Acquired lock in namespace: %s", vmo.Namespace)
 		c.log.Debugf("VMO %s : Spec differences %s", vmo.Name, specDiffs)
 		c.log.Oncef("Updating VMO")
-		metricsExporter.GetSimpleCounterMetrics(metricsExporter.NamesVMOUpdate).Inc()
+		metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesVMOUpdate).Inc()
 		_, err = c.vmoclientset.VerrazzanoV1().VerrazzanoMonitoringInstances(vmo.Namespace).Update(context.TODO(), vmo, metav1.UpdateOptions{})
 		if err != nil {
 			c.log.Errorf("Failed to update status for VMI %s: %v", vmo.Name, err)
@@ -567,7 +567,7 @@ func (c *Controller) syncHandlerStandardMode(vmo *vmcontrollerv1.VerrazzanoMonit
 			c.log.Errorf("Failed to update currentVersion for VMI %s: %v", vmo.Name, err)
 		} else {
 			c.log.Oncef("Updated VMI currentVersion to %s", c.buildVersion)
-			metricsExporter.GetTimestampMetrics(metricsExporter.NamesVMOUpdate).SetLastTime()
+			metricsexporter.GetTimestampMetrics(metricsexporter.NamesVMOUpdate).SetLastTime()
 		}
 	}
 
@@ -602,7 +602,7 @@ func (c *Controller) enqueueVMO(obj interface{}) {
 // IsHealthy returns true if this controller is healthy, false otherwise. It's health is determined based on: (1) its
 // workqueue is 0 or decreasing in a timely manner, (2) it can communicate with API server, and (3) the CRD exists.
 func (c *Controller) IsHealthy() bool {
-	metricsExporter.GetSimpleGaugeMetrics(metricsExporter.NamesQueue).Set(float64(c.workqueue.Len()))
+	metricsexporter.GetSimpleGaugeMetrics(metricsexporter.NamesQueue).Set(float64(c.workqueue.Len()))
 	// Make sure if workqueue > 0, make sure it hasn't remained for longer than 60 seconds.
 	if startQueueLen := c.workqueue.Len(); startQueueLen > 0 {
 		if time.Since(c.lastEnqueue).Seconds() > float64(60) {
