+++ pkg/vmo/deployment.go
@@ -12,7 +12,7 @@ import (
 	vmcontrollerv1 "github.com/verrazzano/verrazzano-monitoring-operator/pkg/apis/vmcontroller/v1"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/config"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/constants"
-	metricsExporter "github.com/verrazzano/verrazzano-monitoring-operator/pkg/metricsexporter"
+	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/metricsexporter"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/resources/deployments"
 	appsv1 "k8s.io/api/apps/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
@@ -42,7 +42,7 @@ func updateOpenSearchDashboardsDeployment(osd *appsv1.Deployment, controller *Co
 		err = updateDeployment(controller, vmo, existingDeployment, osd)
 	}
 	if err != nil {
-		metricsExporter.GetErrorMetrics(metricsExporter.NamesDeploymentUpdateError).Inc()
+		metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError).Inc()
 		controller.log.Errorf("Failed to update deployment %s/%s: %v", osd.Namespace, osd.Name, err)
 		return err
 	}
@@ -52,8 +52,8 @@ func updateOpenSearchDashboardsDeployment(osd *appsv1.Deployment, controller *Co
 
 // CreateDeployments create/update VMO deployment k8s resources
 func CreateDeployments(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitoringInstance, pvcToAdMap map[string]string, existingCluster bool) (dirty bool, err error) {
-	metricsExporter.GetFunctionMetrics(metricsExporter.NamesDeployment).LogStart()
-	defer metricsExporter.GetFunctionMetrics(metricsExporter.NamesDeployment).LogEnd(false) //The error count is incremented by the function which calls createDeployment
+	metricsexporter.GetFunctionMetrics(metricsexporter.NamesDeployment).LogStart()
+	defer metricsexporter.GetFunctionMetrics(metricsexporter.NamesDeployment).LogEnd(false) //The error count is incremented by the function which calls createDeployment
 
 	// Assigning the following spec members seems like a hack; is any
 	// better way to make these values available where the deployments are created?
@@ -96,7 +96,7 @@ func CreateDeployments(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMon
 			}
 		}
 		if err != nil {
-			metricsExporter.GetErrorMetrics(metricsExporter.NamesDeploymentUpdateError).Inc()
+			metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError).Inc()
 			controller.log.Errorf("Failed to update deployment %s/%s: %v", curDeployment.Namespace, curDeployment.Name, err)
 			return false, err
 		}
@@ -144,18 +144,18 @@ func CreateDeployments(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMon
 }
 
 func deleteDeployment(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitoringInstance, deployment *appsv1.Deployment) error {
-	metricsExporter.GetSimpleCounterMetrics(metricsExporter.NamesDeploymentDeleteCounter).Inc()
+	metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentDeleteCounter).Inc()
 	controller.log.Debugf("Deleting deployment %s", deployment.Name)
 	err := controller.kubeclientset.AppsV1().Deployments(vmo.Namespace).Delete(context.TODO(), deployment.Name, metav1.DeleteOptions{})
 	if err != nil {
-		metricsExporter.GetErrorMetrics(metricsExporter.NamesDeploymentDeleteError).Inc()
+		metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentDeleteError).Inc()
 		controller.log.Errorf("Failed to delete deployment %s: %v", deployment.Name, err)
 	}
 	return nil
 }
 
 func updateDeployment(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitoringInstance, existingDeployment, curDeployment *appsv1.Deployment) error {
-	metricsExporter.GetSimpleCounterMetrics(metricsExporter.NamesDeploymentUpdateCounter).Inc()
+	metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter).Inc()
 	var err error
 	curDeployment.Spec.Selector = existingDeployment.Spec.Selector
 	specDiffs := diff.Diff(existingDeployment, curDeployment)
@@ -182,7 +182,7 @@ func rollingUpdate(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitor
 		if !isUpdateAllowed(controller, vmo, current) {
 			continue
 		}
-		metricsExporter.GetSimpleCounterMetrics(metricsExporter.NamesDeploymentUpdateCounter).Inc()
+		metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter).Inc()
 		// Selector may not change, so we copy over from existing
 		current.Spec.Selector = existing.Spec.Selector
 		// Deployment spec differences, so call Update() and return
@@ -192,7 +192,7 @@ func rollingUpdate(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitor
 			controller.log.Oncef("Updating deployment %s in namespace %s", current.Name, current.Namespace)
 			_, err = controller.kubeclientset.AppsV1().Deployments(vmo.Namespace).Update(context.TODO(), current, metav1.UpdateOptions{})
 			if err != nil {
-				metricsExporter.GetErrorMetrics(metricsExporter.NamesDeploymentUpdateError).Inc()
+				metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError).Inc()
 				return false, err
 			}
 			//okay to return dirty=false after updating the *last* deployment
@@ -222,10 +222,10 @@ func updateAllDeployments(controller *Controller, vmo *vmcontrollerv1.Verrazzano
 		if err != nil {
 			return false, err
 		}
-		metricsExporter.GetSimpleCounterMetrics(metricsExporter.NamesDeploymentUpdateCounter).Inc()
+		metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter).Inc()
 		_, err = controller.kubeclientset.AppsV1().Deployments(vmo.Namespace).Update(context.TODO(), curDeployment, metav1.UpdateOptions{})
 		if err != nil {
-			metricsExporter.GetErrorMetrics(metricsExporter.NamesDeploymentUpdateError).Inc()
+			metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError).Inc()
 			return false, err
 		}
 	}
