+++ pkg/vmo/ingress.go
@@ -10,7 +10,7 @@ import (
 	"github.com/verrazzano/pkg/diff"
 	vmcontrollerv1 "github.com/verrazzano/verrazzano-monitoring-operator/pkg/apis/vmcontroller/v1"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/constants"
-	metricsExporter "github.com/verrazzano/verrazzano-monitoring-operator/pkg/metricsexporter"
+	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/metricsexporter"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/resources/ingresses"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -20,12 +20,12 @@ import (
 
 // CreateIngresses create/update VMO ingress k8s resources
 func CreateIngresses(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitoringInstance) error {
-	metricsExporter.GetFunctionMetrics(metricsExporter.NamesIngress).LogStart()
-	defer metricsExporter.GetFunctionMetrics(metricsExporter.NamesIngress).LogEnd(false)
+	metricsexporter.GetFunctionMetrics(metricsexporter.NamesIngress).LogStart()
+	defer metricsexporter.GetFunctionMetrics(metricsexporter.NamesIngress).LogEnd(false)
 	ingList, err := ingresses.New(vmo)
 	if err != nil {
 		controller.log.Errorf("Failed to create Ingress specs for VMI %s: %v", vmo.Name, err)
-		metricsExporter.GetFunctionMetrics(metricsExporter.NamesIngress).IncError()
+		metricsexporter.GetFunctionMetrics(metricsexporter.NamesIngress).IncError()
 		return err
 	}
 	if vmo.Spec.IngressTargetDNSName == "" {
@@ -42,7 +42,7 @@ func CreateIngresses(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonit
 			// resource otherwise. Instead, the next time the resource is updated
 			// the resource will be queued again.
 			runtime.HandleError(errors.New("ingress name must be specified"))
-			metricsExporter.GetFunctionMetrics(metricsExporter.NamesIngress).IncError()
+			metricsexporter.GetFunctionMetrics(metricsexporter.NamesIngress).IncError()
 			return nil
 		}
 
@@ -58,13 +58,13 @@ func CreateIngresses(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonit
 			_, err = controller.kubeclientset.NetworkingV1().Ingresses(vmo.Namespace).Create(context.TODO(), curIngress, metav1.CreateOptions{})
 		} else {
 			controller.log.Errorf("Failed getting existing Ingress %s/%s: %v", vmo.Namespace, ingName, err)
-			metricsExporter.GetFunctionMetrics(metricsExporter.NamesIngress).IncError()
+			metricsexporter.GetFunctionMetrics(metricsexporter.NamesIngress).IncError()
 			return err
 		}
 
 		if err != nil {
 			controller.log.Errorf("Failed to create/update Ingress %s/%s: %v", vmo.Namespace, ingName, err)
-			metricsExporter.GetFunctionMetrics(metricsExporter.NamesIngress).IncError()
+			metricsexporter.GetFunctionMetrics(metricsexporter.NamesIngress).IncError()
 			return err
 		}
 	}
@@ -74,7 +74,7 @@ func CreateIngresses(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonit
 	selector := labels.SelectorFromSet(map[string]string{constants.VMOLabel: vmo.Name})
 	existingIngressList, err := controller.ingressLister.Ingresses(vmo.Namespace).List(selector)
 	if err != nil {
-		metricsExporter.GetFunctionMetrics(metricsExporter.NamesIngress).IncError()
+		metricsexporter.GetFunctionMetrics(metricsexporter.NamesIngress).IncError()
 		return err
 	}
 	for _, ingress := range existingIngressList {
@@ -85,7 +85,7 @@ func CreateIngresses(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonit
 				controller.log.Errorf("Failed to delete Ingress %s/%s: %v", vmo.Namespace, ingress.Name, err)
 				return err
 			}
-			metricsExporter.GetSimpleCounterMetrics(metricsExporter.NamesIngressDeleted).Inc()
+			metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesIngressDeleted).Inc()
 		}
 	}
