+++ pkg/vmo/rolebindings.go
@@ -10,7 +10,7 @@ import (
 	"github.com/verrazzano/pkg/diff"
 	vmcontrollerv1 "github.com/verrazzano/verrazzano-monitoring-operator/pkg/apis/vmcontroller/v1"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/constants"
-	metricsExporter "github.com/verrazzano/verrazzano-monitoring-operator/pkg/metricsexporter"
+	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/metricsexporter"
 	"github.com/verrazzano/verrazzano-monitoring-operator/pkg/resources"
 	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -19,7 +19,7 @@ import (
 
 // CreateRoleBindings creates/updates VMO RoleBindings k8s resources
 func CreateRoleBindings(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitoringInstance) error {
-	metricsExporter.GetSimpleCounterMetrics(metricsExporter.NamesRoleBindings).Inc()
+	metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesRoleBindings).Inc()
 	controller.log.Oncef("Creating/updating RoleBindings for VMI %s", vmo.Name)
 
 	newRoleBindings, err := NewRoleBindings(vmo, controller)
@@ -76,7 +76,7 @@ func CreateRoleBindings(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMo
 			}
 		}
 	}
-	metricsExporter.GetTimestampMetrics(metricsExporter.NamesRoleBindings).SetLastTime()
+	metricsexporter.GetTimestampMetrics(metricsexporter.NamesRoleBindings).SetLastTime()
 	return nil
 }
